+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added the (missing) 'prismatic oil' icon
 - Changed the Settings Menu Login url from '/login/email' to '/login' and added a custom polyfill to make the login page buttons work
 - Replaced the 'online/offline' trade status option with more options related to 'instant buyout' and 'in person' trades
+- Fixed exchange rate retrieval for Reliquary Keys
 
 ## 0.8.40 (2026-03-13)